requirements.txt
python-decouple==3.3
requests==2.22.0